// lib/api/admin.ts
export interface SessionEBMInfo {
  ebmTin: string | null
  purchaseCode: string | null
  customerPhone: string | null
  ebmPaymentMethodCode: string | null
  ebmPaymentMethodName: string | null
}

export interface UpdateSessionEBMInfoData {
  ebmTin?: string | null
  purchaseCode?: string | null
  customerPhone?: string | null
  ebmPaymentMethodCode?: string | null
  ebmPaymentMethodName?: string | null
}
// ...
export interface VsdcCodeItem {
  code: string
  name: string
}